+++ lib/page/detailPage.dart
